% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubert_t.R
\name{deltaSub}
\alias{deltaSub}
\title{Utility function}
\usage{
deltaSub(i, j, group_list)
}
\arguments{
\item{i}{First index}

\item{j}{Second index}

\item{group_list}{List of indices for each block}
}
\description{
This function outputs 1 if i and j are in at least one group together, 
and 0 otherwise
}
