% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubert_t.R
\name{makeGroupList}
\alias{makeGroupList}
\title{Convert character string in lavaan syntax into a list of indices}
\usage{
makeGroupList(groups, A)
}
\arguments{
\item{groups}{Character string in lavaan syntax specifying groups}

\item{A}{A Distance or similarity matrix. Must have column names}
}
\value{
group_list List of column indices of A corresponding to each group
}
\description{
This sub-routine is called by \code{matrixStrucTest} and \code{prepBoxPlots}.
}
