% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubert_t.R
\name{matrixTest}
\alias{matrixTest}
\title{Gamma and t function}
\usage{
matrixTest(A, group_list, B = 1000, absolute = TRUE)
}
\arguments{
\item{A}{Distance or similarity matrix, e.g. correlation}

\item{group_list}{List of groupings}

\item{B}{Number of Monte Carlo resamples (defaults to B=1000)}

\item{absolute}{Use the absolute values of A (defaults to TRUE)}
}
\value{
pt_overall_one_sided: Overall one-sided p-value using t statistic

pt_overall_two_sided: Overall two-sided p-value using t statistic

pt_multi_one_sided: Block-specific one-sided p-values using t statistic

pt_multi_two_sided: Block-specific two-sided p-values using t statistic

t0 Observed overall: t statistic

t0k: Observed block-specific t statistic

t_overall: Vector of overall t statistics from permuted A

t_max_one_sided: Vector of max t statistics from permuted A (one-sided)

t_max_two_sided: Vector of max t statistics from permuted A (two-sided)

pG_overall_one_sided: Overall one-sided p-value using Hubert's Gamma

pG_overall_two_sided: Overall two-sided p-value using Hubert's Gamma

pG_multi_one_sided: Block-specific one-sided p-values using Hubert's Gamma

pG_multi_two_sided: Block-specific two-sided p-values using Hubert's Gamma

Gamma0: Observed overall Hubert's Gamma

Gamma0k: Observed block-specific Hubert's Gamma

Gamma_overall: Vector of Hubert's Gamma statistics from permuted A

Gamma_max_one_sided: Vector of max Hubert's Gamma statistics from permuted A (one-sided)

Gamma_max_two_sided: Vector of max Hubert's Gamma statistics from permuted A (two-sided)

B: number of Monte Carlo resamples
}
\description{
This function computes p-values for Hubert's Gamma and t statistics
}
\examples{
# example for matrixText package
library(matrixTest)
data("big5")

# get column numbers for questionnaire items
items <- grep("[0-9]", colnames(big5))

# compute Spearman's correlation matrix
A <- cor(big5[, items], use = "complete.obs", method = "spearman")

# get column numbers for items within each block/group
extrovert <- grep("E", colnames(A))
neurotic <- grep("N", colnames(A))
agreeable <- grep("A", colnames(A))
conscientious <- grep("C", colnames(A))
open <- grep("O", colnames(A))

# put blocks/groups in list for matrixTest function
group_list <- list(extrovert = extrovert, 
                   neurotic = neurotic, 
                   agreeable = agreeable, 
                   conscientious = conscientious,
                   open = open)

# compute permutation p-values
# Note: two-sided p-values from Hubert's Gamma printed by default
#       other results available by directing accessing them from the
#       returned object
result <- matrixTest(A = A, group_list = group_list, B = 1000, absolute = TRUE)
result

# Visualize groups
library(ggplot2)
library(reshape2)

ord <- unlist(group_list)
diag(A) <- NA # remove diagonals from color scale
Am <- melt(A[ord, ord])
names(Am) <- c("x", "y", "value")
Am$y <- factor(Am$y, levels = rev(levels(Am$y)))

ggplot(aes(x = x, y = y, fill = abs(value)), data = Am)+
  geom_tile()+
  theme_bw(18)+
  scale_fill_gradient2(space="Lab", name="abs(Cor)", lim = c(0, 1))+
  labs(x = "", y = "")+
  theme(axis.text.x = element_text(angle = 90, vjust = .35,hjust=1))
}
\keyword{Gamma,}
\keyword{Hubert's}
\keyword{MC}
\keyword{permutation,}
\keyword{resampling}
\keyword{t-statistic,}

