% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubert_t.R
\name{multiSub}
\alias{multiSub}
\title{Utility function}
\usage{
multiSub(i, j, group)
}
\arguments{
\item{i}{first index}

\item{j}{second index}

\item{group_list}{list of blocks}
}
\description{
This function outputs TRUE if either i or j are in group, FALSE otherwise
}
