% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubert_t.R
\name{print.mst}
\alias{print.mst}
\title{Print results from \code{matrixStrucTest}}
\usage{
\method{print}{mst}(x, ...)
}
\arguments{
\item{x}{Output from matrixStrucTest}

\item{...}{Further arguments passed to print}
}
\description{
This function prints results from an object returned by \code{matrixStrucTest}.
}
\keyword{matrixStrucTest}
\keyword{print}
