% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubert_t.R
\name{print.mt}
\alias{print.mt}
\title{Print function for results}
\usage{
\method{print}{mt}(mt)
}
\arguments{
\item{mt}{Output from matrixTest}
}
\description{
This function prints results from matrixTest
}
\keyword{matrixTest}
\keyword{print}
