% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_filon.R
\name{sigmaRhoFun}
\alias{sigmaRhoFun}
\title{Utility function}
\usage{
sigmaRhoFun(j, k, h, m, A)
}
\arguments{
\item{j}{First index}

\item{k}{Second index}

\item{h}{Third index}

\item{m}{Fourth index}

\item{A}{Correlation matrix}
}
\description{
Utility function
}
