% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubert_t.R
\name{deltaSub}
\alias{deltaSub}
\title{Utility function}
\usage{
deltaSub(i, j, group_list)
}
\arguments{
\item{i}{first index}

\item{j}{second index}

\item{group_list}{list of blocks}
}
\description{
This function outputs 1 if i and j are in at least one group together, 
and 0 otherwise
}

