% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubert_t.R
\name{deltaSub}
\alias{deltaSub}
\title{Sub-routine to create Delta matrix}
\usage{
deltaSub(i, j, group_list)
}
\arguments{
\item{i}{First index}

\item{j}{Second index}

\item{group_list}{List of indices for each block}
}
\description{
This sub-routine outputs 1 if i and j are in at least one group together, 
and 0 otherwise, and is called by \code{matrixStrucTest} and 
\code{prepBoxPlots}.
}
