% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubert_t.R
\name{multiSub}
\alias{multiSub}
\title{Utility function}
\usage{
multiSub(i, j, group)
}
\arguments{
\item{i}{First index}

\item{j}{Second index}

\item{group}{Indices for items in group}
}
\description{
This function outputs TRUE if either i or j are in group, FALSE otherwise
}
