% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubert_t.R
\name{multiSub}
\alias{multiSub}
\title{Sub-routine to create Delta matrix for block-specific tests}
\usage{
multiSub(i, j, group)
}
\arguments{
\item{i}{First index}

\item{j}{Second index}

\item{group}{Indices for items in group}
}
\description{
This sub-routine outputs TRUE if either \code{i} or \code{j} are in 
\code{group}, FALSE otherwise, and is called by \code{matrixStrucTest} 
and \code{prepBoxPlots}.
}
