% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubert_t.R
\name{prepBoxPlots}
\alias{prepBoxPlots}
\title{Prepare data for box plots}
\usage{
prepBoxPlots(A, group_list, absolute = TRUE)
}
\arguments{
\item{A}{Distance or similarity matrix, e.g. correlation}

\item{group_list}{List of groupings}

\item{absolute}{Use the absolute values of A (defaults to TRUE)}

\item{B}{Number of Monte Carlo resamples (defaults to B=1000)}
}
\value{
multi: data frame for making box plots for block-specific tests

overall: data frame for making box plots for overall test
}
\description{
This function prepares the data for making box plots
}
\examples{
library(matrixTest)
library(ggplot2)

data("big5")

# get column numbers for questionnaire items
items <- grep("[0-9]", colnames(big5))

# compute Spearman's correlation matrix
A <- cor(big5[, items], use = "complete.obs", method = "spearman")

# get column numbers for items within each block/group
extrovert <- grep("E", colnames(A))
neurotic <- grep("N", colnames(A))
agreeable <- grep("A", colnames(A))
conscientious <- grep("C", colnames(A))
open <- grep("O", colnames(A))

# put blocks/groups in list for matrixTest function
group_list <- list(extrovert = extrovert, 
                   neurotic = neurotic, 
                   agreeable = agreeable, 
                   conscientious = conscientious,
                   open = open)

# Make box plots contrasting within and between group correlations
box <- prepBoxPlots(A = A, group_list, absolute = TRUE)

ggplot(aes(x = as.factor(delta), y = a), data = box$overall)+
  geom_boxplot()+
  theme_bw(22)+
  labs(x = expression(Delta), y="|a|")

dev.new(width = 12, height = 5)
ggplot(aes(x = as.factor(delta), y = a), data = box$multi)+
  geom_boxplot()+
  facet_grid(~block)+
  theme_bw(22)+
labs(x = expression(Delta), y = "|a|")
}

