% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_filon.R
\name{sigmaRhoFun}
\alias{sigmaRhoFun}
\title{Sub-routine to compute OLS estimates of covariance between correlations}
\usage{
sigmaRhoFun(j, k, h, m, A)
}
\arguments{
\item{j}{First index}

\item{k}{Second index}

\item{h}{Third index}

\item{m}{Fourth index}

\item{A}{Correlation matrix}
}
\description{
This sub-routine is called by \code{X2fun}.
}
