% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_filon.R
\name{sigmaZFun}
\alias{sigmaZFun}
\title{Utility function}
\usage{
sigmaZFun(s, t, index, A, Sigma)
}
\arguments{
\item{s}{First index}

\item{t}{Second index}

\item{index}{Matrix with two columns with index pairs given by rows}

\item{A}{Correlation matrix}

\item{Sigma}{Variance-covariances of correlation matrix A}
}
\description{
Utility function
}
